fix : 모든 PR 코멘트 해결 및 코드 수정 완료
- 소스파일 분리
- 불필요한 ValueNotifier 상태 전달
- 음성 인식 활성 UI로 변경할 때 자연스러운 트랜지션 적용
- 상태 조작 위치 변경
- 메소드 argument 최소화
- useMemoized 사용 이유 설명
- 자연스러운 트래진션 적용
